requirements.txt
boto3==1.17.7
Pillow==8.1.0
